Moved U-SQL helper methods out of Job Manager to be testable
